User provider clean up
